refactor(LottoResult): 로또 결과를 출력하는 로직을 분리해 새로운 클래스로 추가
LottoResult 클래스는 다음과 같은 메서드와 생성자를 가집니다.
- 당첨 결과를 저장해두는 result
- 당첨 결과를 출력하기 위한 모든 메서드를 실행시키는 print 메서드
- 당첨 번호와 로또의 일치 여부를 계산하는 calculateMatching 메서드
- 일치하는 번호 갯수를 세는 countMatching 메서드
- 보너스 번호가 포함되어 있는지 검사하는 checkBonus 메서드
- 수익률을 계산하는 calculateRate 메서드
- 당첨 결과를 출력하는 printResult 메서드
