+++ docs/README.md
@@ -44,7 +44,7 @@
 리펙토링
 
 - [x] 로또를 발행하는 로직 클래스화
-- [ ] 로또 당첨 통계 계산 및 출력 로직 클래스화
+- [x] 로또 당첨 통계 계산 및 출력 로직 클래스화
 
 ---
